+++ src/components/map/airports/MapAirportsList.vue
@@ -100,7 +100,9 @@ const url = new URL(location.href);
 if (url.searchParams.has('start') && url.searchParams.has('end')) {
     start.setTime(Number(url.searchParams.get('start')));
     end.setTime(Number(url.searchParams.get('end')));
-    store.mapSettings.bookingOverride = true;
+    setUserMapSettings({
+        bookingOverride: true,
+    });
 }
 
 const { data } = await useAsyncData('bookings', async () => {
@@ -325,8 +327,6 @@ watch(map, val => {
             wrapX: false,
         });
 
-        const styles = airportLayoutStyles();
-
         gatesLayer = new VectorLayer<any>({
             source: gatesSource.value,
             properties: {
@@ -354,6 +354,11 @@ watch(map, val => {
     val.on('click', handleMapClick);
 }, {
     immediate: true,
+});
+
+watch(dataStore.vatsim.data.bars, val => {
+    if (!Object.keys(val).length) return;
+    airportLayerSource.value?.dispatchEvent('change');
 });
 
 onBeforeUnmount(() => {