+++ src/components/views/header/ViewHeader.vue
@@ -25,7 +25,10 @@
             >
                 <common-logo/>
             </nuxt-link>
-            <div class="header__sections __from-tablet">
+            <div
+                v-if="store.viewport.width > 1000"
+                class="header__sections"
+            >
                 <div class="header__sections_section header__buttons">
                     <div
                         v-for="button in headerMenu"
@@ -172,26 +175,72 @@
                         <github-icon/>
                     </template>
                 </common-button>
-                <common-button
-                    href="https://docs.vatsim-radar.com"
-                    size="S"
-                    target="_blank"
-                    type="secondary"
+                <common-tooltip
+                    location="bottom"
+                    width="max-content"
                 >
-                    <template #icon>
-                        <question-icon/>
+                    <template #activator>
+                        <common-button
+                            href="https://docs.vatsim-radar.com"
+                            size="S"
+                            target="_blank"
+                            type="secondary"
+                        >
+                            <template #icon>
+                                <question-icon/>
+                            </template>
+                        </common-button>
                     </template>
-                </common-button>
-                <common-button
-                    href="/discord"
-                    size="S"
-                    target="_blank"
-                    type="secondary"
+                    Documentation
+                </common-tooltip>
+                <common-tooltip
+                    v-if="app.$pwa && !app.$pwa?.isPWAInstalled && app.$pwa?.showInstallPrompt"
+                    location="bottom"
+                    width="max-content"
                 >
-                    <template #icon>
-                        <discord-icon/>
+                    <template #activator>
+                        <common-button
+                            size="S"
+                            type="secondary"
+                            @click="app.$pwa?.install()"
+                        >
+                            <template #icon>
+                                <load-on-pc-icon/>
+                            </template>
+                        </common-button>
                     </template>
-                </common-button>
+                    Install App
+                </common-tooltip>
+                <common-tooltip
+                    align="right"
+                    location="bottom"
+                    width="200px"
+                >
+                    <template #activator>
+                        <common-button
+                            size="S"
+                            type="secondary"
+                        >
+                            <template #icon>
+                                <discord-icon/>
+                            </template>
+                        </common-button>
+                    </template>
+                    <div class="__info-sections">
+                        <common-button
+                            href="https://discord.com/invite/vatsim"
+                            target="_blank"
+                        >
+                            General VATSIM Discussion
+                        </common-button>
+                        <common-button
+                            href="/discord"
+                            target="_blank"
+                        >
+                            VATSIM Radar Development
+                        </common-button>
+                    </div>
+                </common-tooltip>
                 <common-button
                     v-if="config.public.IS_DOWN !== 'true'"
                     size="S"
@@ -257,6 +306,8 @@ import ViewHeaderMobileMenu from '~/components/views/header/ViewHeaderMobileMenu
 import ViewHeaderThemeSwitcher from '~/components/views/header/ViewHeaderThemeSwitcher.vue';
 import StarFilledIcon from '@/assets/icons/kit/star-filled.svg?component';
 import CommonBubble from '~/components/common/basic/CommonBubble.vue';
+import LoadOnPcIcon from '~/assets/icons/kit/load-on-pc.svg?component';
+import CommonTooltip from '~/components/common/basic/CommonTooltip.vue';
 
 const headerMenu = useHeaderMenu();
 
@@ -269,6 +320,8 @@ const notamCookie = useCookie<boolean>('notam-closed', {
     secure: true,
     maxAge: 60 * 60 * 24 * 7,
 });
+
+const app = useNuxtApp();
 
 const isMobileOrTablet = useIsMobileOrTablet();
 const isMobile = useIsMobile();