+++ src/utils/backend/tasks.ts
@@ -7,7 +7,7 @@ import type {
     PatreonPledgesReward,
     PatreonPledgesUser,
 } from '~/types/data/patreon';
-import { radarStorage } from '~/utils/backend/storage';
+import { isDataReady, radarStorage } from '~/utils/backend/storage';
 import { initNavigraph } from '~/utils/backend/navigraph-db';
 import { updateSimAware } from '~/utils/backend/vatsim/simaware';
 import { updateVatglassesData } from '~/utils/backend/vatglasses';
@@ -30,9 +30,14 @@ function basicTasks() {
     defineCronJob('15 */2 * * *', updateVatglassesData).catch(console.error);
     defineCronJob('15 * * * *', updateVatSpy).catch(console.error);
 
-    redisSubscriber.subscribe('vatglassesActive');
-    redisSubscriber.on('message', (_, message) => {
-        radarStorage.vatglasses.activeData = message;
+    redisSubscriber.subscribe('vatglassesActive', 'vatglassesDynamic');
+    redisSubscriber.on('message', (channel, message) => {
+        if (channel === 'vatglassesActive') {
+            radarStorage.vatglasses.activeData = message;
+        }
+        else if (channel === 'vatglassesDynamic') {
+            radarStorage.vatglasses.dynamicData = JSON.parse(message);
+        }
     });
 }
 
@@ -215,7 +220,7 @@ export async function setupRedisDataFetch() {
     await defineCronJob('15 * * * *', async () => {
         await updateData();
 
-        while (!radarStorage.vatspy.data) {
+        while (!await isDataReady()) {
             await sleep(1000 * 60);
             await updateData();
         }