+++ src/utils/data/vatglasses.ts
@@ -28,6 +28,7 @@ let facilities: {
 /*
 If we want to get the initial load of the combined sectors, we need sectorsCombined, airspaceKeys and the vatglasses Data version of the server on which base the sectors were combined, to make sure we have the same data version on the client side.
 */
+const VGdebugMode = false;
 
 export interface VatglassesActiveData {
     vatglassesActiveRunways: VatglassesActiveRunways;
@@ -81,6 +82,7 @@ function updateVatglassesPositionsAndAirspaces() {
     const vatglassesData = dataStore?.vatglasses?.value?.data ?? radarStorage?.vatglasses?.data.data;
     const vatglassesActiveRunways = dataStore?.vatglassesActiveRunways.value ?? workerDataStore.vatglassesActiveRunways;
     const vatglassesActivePositions = dataStore?.vatglassesActivePositions.value ?? workerDataStore.vatglassesActivePositions;
+    const vatglassesDynamicData = dataStore?.vatglassesDynamicData?.value ?? radarStorage.vatglasses.dynamicData;
     const vatsimData = dataStore?.vatsim ?? radarStorage.vatsim;
     if (!vatglassesData || !vatsimData) return newVatglassesActivePositions;
 
@@ -111,7 +113,7 @@ function updateVatglassesPositionsAndAirspaces() {
         }
 
         // TODO: sort firs by the same as EuroScope sorts the positions before assigning logged in stations to a position
-        // let first = 0; // used for debug
+        let first = 0; // used for debug
         const firs = dataStore?.vatsim?.data?.firs.value ?? radarStorage?.vatsim?.firs;
         for (const fir of [...firs, ...arrivalController]) { // this has an entry for each center controller connected. const fir is basically a center controller
             // In data.firs it is called controller, in data.locals it is called atc, so we have to get the correct one
@@ -125,23 +127,24 @@ function updateVatglassesPositionsAndAirspaces() {
             if (!atc) continue;
 
             // // used for debug
-            // if (first === 0) {
-            //     atc.frequency = '135.050';
-            //     atc.callsign = 'AFRE_FSS';
-            //     atc.cid = 1025793;
-            //     first++;
-            // }
-            // else if (first === 1) {
-            //     break;
-            //     if (calls > 3) break;
-            //     atc.frequency = '129.100';
-            //     atc.callsign = 'EDMM_ALB_CTR';
-            //     atc.cid = 1025794;
-            // // first++;
-            // }
-            // else {
-            //     break;
-            // }
+            if (VGdebugMode) {
+                if (first === 0) {
+                    atc.frequency = '121.550';
+                    atc.callsign = 'ENOR_S_CTR';
+                    atc.cid = 1025793;
+                    first++;
+                }
+                else if (first === 1) {
+                    break;
+                    atc.frequency = '127.755';
+                    atc.callsign = 'ESMM_2_CTR';
+                    atc.cid = 1025794;
+                    // first++;
+                }
+                else {
+                    break;
+                }
+            }
 
             let foundMatchingVatglassesController = false;
             let doublePositionMatch = false;
@@ -157,7 +160,7 @@ function updateVatglassesPositionsAndAirspaces() {
                 const countryGroup = vatglassesData[countryGroupId];
                 for (const vatglassesPositionId in countryGroup.positions) {
                     const vatglassesPosition = countryGroup.positions[vatglassesPositionId];
-                    if (vatglassesPosition.frequency !== atc.frequency) continue;
+                    if (vatglassesPosition.frequency && vatglassesPosition.frequency !== atc.frequency) continue;
                     if (!atc.callsign.endsWith(vatglassesPosition.type)) continue;
                     if (!vatglassesPosition.pre.some((prefix: string) => atc.callsign.startsWith(prefix))) continue;
                     if (vatglassesActiveController?.[countryGroupId]?.[vatglassesPositionId] != null) {
@@ -182,7 +185,6 @@ function updateVatglassesPositionsAndAirspaces() {
 
     // TODO: We could compare the entries of `vatglassesActiveStations` with `dataStore.vatglassesActivePositions.value[countryGroupId][positionId]`, and if they are equal, no position has changed and we have nothing to do. However, we need to consider if a runway was updated or changed in the frontend and therefore this update cycle was called. Maybe use a parameter in this function like `runwayChanged`, and when we call updateVatglassesState() in the `activeRunwayChanged` function, we pass this parameter.
 
-
     // Fill the vatglassesActiveAirspaces object with the active airspaces
     vatglassesActiveAirspaces = {};
     for (const countryGroupId in vatglassesData) {
@@ -190,8 +192,17 @@ function updateVatglassesPositionsAndAirspaces() {
 
         let activeGroupVatglassesPosition: string[] = [];
         if (vatglassesActiveController[countryGroupId]) activeGroupVatglassesPosition = Object.keys(vatglassesActiveController[countryGroupId]);
-        for (const [airspaceIndex, airspace] of countryGroup.airspace.entries()) {
-            const vatglassesPositionId = airspace.owner.find((element: string) => {
+        for (const airspaceIndex in countryGroup.airspace) {
+            const airspace = countryGroup.airspace[airspaceIndex];
+            let airspaceOwner: string [] = [];
+            if (vatglassesDynamicData?.data?.[countryGroupId]?.airspace?.[airspaceIndex]) {
+                airspaceOwner = vatglassesDynamicData?.data?.[countryGroupId]?.airspace?.[airspaceIndex];
+            }
+            else if (airspace?.owner) {
+                airspaceOwner = airspace.owner;
+            }
+
+            const vatglassesPositionId = airspaceOwner.find((element: string) => {
                 if (element.includes('/')) { // Covers this cases: Positions from other data files can be referenced in the format country/position, where position is defined in country.json
                     const elementSplit = element.split('/');
                     const otherCountryCode = elementSplit[0];
@@ -651,7 +662,7 @@ async function initVatglassesCombined() {
     combineDataInitialized = true;
     dataStore.vatglassesCombiningInProgress.value = true;
     try {
-        const data: VatglassesActiveData = JSON.parse(await $fetch<string>(`/api/data/vatsim/data/vatglasses-active`));
+        const data: VatglassesActiveData = JSON.parse(await $fetch<string>(`/api/data/vatsim/data/vatglasses/active`));
         const vatglassesDataVersion = dataStore?.vatglasses?.value?.version;
         if (vatglassesDataVersion === data.version) {
             for (const countryGroupId in data.vatglassesActivePositions) {
@@ -684,7 +695,7 @@ async function initVatglassesCombined() {
         const localRunways = dataStore.vatglassesActiveRunways.value;
 
         for (const icao in localRunways) {
-            if (localRunways[icao].active !== serverRunways?.[icao].active) {
+            if (localRunways[icao].active !== serverRunways?.[icao]?.active) {
                 await activeRunwayChanged(icao, false);
             }
         }
@@ -727,7 +738,7 @@ export async function initVatglasses(inputMode: string = 'local', serverDataStor
             await initVatglassesCombined();
         }
 
-        watch([dataStore.vatsim.data.firs, dataStore.vatsim.data.locals, vatglassesCombined], async () => {
+        watch([dataStore.vatsim.data.firs, dataStore.vatsim.data.locals, dataStore.vatglassesDynamicData, vatglassesCombined], async () => {
             if (!combineDataInitialized && vatglassesCombined.value) {
                 await updateVatglassesStateLocal(true);
                 await initVatglassesCombined();